refactor: 테스트용 클래스 패키지 이동
